fix: allow custom currency units to specify derivation path
You can't currently run a mint using a custom currency unit
because it throws Error::UnsupportedUnit when getting the
derivation path in Mint::new.

This fix allows the creator of a custom CurrencyUnit to specify
the derivation path.

It formats these custom currency units into strings like UNIT:PATH
where UNIT is the currency name string and PATH is the derivation
path int.