Custom pool
Implement a simple connection pool to manage multiple connections to SQLite.

The goal is to reduce external dependencies, even with popular crates such as
r2d2, to minimize the vendor-supply attack vector.

If other parts of the code require a generic connection pool, this module can
be promoted to be an independent crate.