Introduce a generic pubsub mod in cdk-common
Remove all previous implementations that were cumbersome and complicated, full
of generics, in favor of a simpler generic pubsub with simpler concepts.

The pubsub has a simple database optional implementation to persist the latest
messages and send them to new subscribers.

The same codebase is shared between the mint and the wallet. The wallet
implements a remote pubsub subscription, which connects to a remote pubsub via
two methods: polling or long connections, both of which are abstracted.
Internally, a local pubsub is used to distribute messages efficiently and
coalesce subscriptions

TODO: Implement the remote subscription into the wallet