Undo all the over-engineered solutions I proposed to try to adapt to the workload (mostly updates and rarely inserts).
I rolled back the proposed solution to the initial INSERT, which will most
likely fail and fall back to an UPDATE, but in exchange, the code is much
simpler to debug and maintain. If this approach proves to be a bottleneck in
the future, I will bring back 3ad85294898979508a218b3e03a71ad41cc00a2d, along
with mechanisms to notify the mints whenever the Signatory has new keys.