Fixed race conditions
The first time keys are fetched from the mint, they do not yet exist in the
wallet database; therefore, the counter is not updated correctly, because the
code performs an update that expects the keyset row to exist.

To solve this issue, I introduced a new keyset_counter table that will upsert
the data, even if the proper keyset info is pulled later.